requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
rich>=13.0.0
feedparser>=6.0.0
openai>=1.0.0
numpy>=1.24.0
